#14 Register fix

+++ src/main/java/org/forum/controller/SectionController.java
@@ -12,10 +12,7 @@ import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.validation.BindingResult;
-import org.springframework.web.bind.annotation.ModelAttribute;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestMethod;
+import org.springframework.web.bind.annotation.*;
 import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
 import javax.validation.Valid;
@@ -34,6 +31,7 @@ public class SectionController {
     @Autowired
     private UserService userService;
 
+    /** SECTION OVERVIEW **/
     @RequestMapping("{id}")
     public String getTopicsFromSection(@PathVariable int id,
                                        Model model) {
@@ -42,37 +40,40 @@ public class SectionController {
         return "section";
     }
 
+    /** SECTION MODERATORS OVERVIEW **/
     @RequestMapping(value = "{id}/moderators", method = RequestMethod.GET)
     public String getModeratorsOfSection(@PathVariable int id,
                                        Model model) {
         model.addAttribute("section", sectionService.findOne(id));
         model.addAttribute("users", sectionService.findOne(id).getModerators());
-        return "moderators_or_members";
+        return "moderators_and_members_view";
     }
 
+
+    /** SECTION MEMBERS OVERVIEW **/
     @RequestMapping(value = "{id}/members", method = RequestMethod.GET)
     public String getMembersOfSection(@PathVariable int id, Model model) {
         model.addAttribute("section", sectionService.findOne(id));
         model.addAttribute("users", sectionService.findOne(id).getMembers());
-        return "moderators_or_members";
+        return "moderators_and_members_view";
     }
 
+    /** SECTION MEMBERS OR MODERATORS ADDING VIEW **/
     @RequestMapping(value = "{id}/add", method = RequestMethod.GET)
-    public String addModeratorsAndMembersToSection(@PathVariable int id, Model model){
+    public String showUsersToAdd(@PathVariable int id, Model model){
         model.addAttribute("section", sectionService.findOne(id));
         model.addAttribute("users", userService.findAll());
         return "add";
     }
 
-
-
-
+    /** NEW SECTION OVERVIEW GET METHOD**/
     @RequestMapping(value = "new", method = RequestMethod.GET)
     public String getNewSectionForm(Model model) {
         model.addAttribute("newSection", new NewSectionForm());
         return "new_section_form";
     }
 
+    /** CREATE NEW SECTION AND REDIRECT TO THIS SECTION VIEW**/
     @RequestMapping(value = "new", method = RequestMethod.POST)
     public String processAndAddNewSection(
             @Valid @ModelAttribute("newSection") NewSectionForm newSkupina,
@@ -96,15 +97,13 @@ public class SectionController {
             section.setIsPublic(0);
         }
         section.setUser(user);
+        section.getModerators().add(user);
         section = sectionService.save(section);
         return "redirect:/section/" + section.getId();
     }
 
-//    @RequestMapping(value = "moderators", method = RequestMethod.GET)
-//    public String showModeratorsOfSections(Model model){
-//        model.addAttribute("sections", sectionService.)
-//    }
 
+    /** DELETE SECTION **/
     @RequestMapping(value = "delete/{id}", method = RequestMethod.GET)
     public String delete(@PathVariable int id,
                          Authentication authentication,
@@ -114,6 +113,11 @@ public class SectionController {
 //        if (!user.getRoles().contains(adminRole)) {
 //            return "redirect:/section/" + id;
 //        }
+        Section section = sectionService.findOne(id);
+        if(!user.equals(section.getUser()) || !section.getModerators().contains(user)){
+            return "redirect:/section/" + id;
+        }
+
         sectionService.delete(id);
 
         model.addFlashAttribute("message", "section.successfully.deleted");
